+++ src/minicraft/core/Game.java
@@ -32,7 +32,7 @@ public class Game {
 	public static boolean HAS_GUI = true;
 	
 	public static final String NAME = "Minicraft Plus"; // This is the name on the application window
-	public static final Version VERSION = new Version("2.0.6-dev1");
+	public static final Version VERSION = new Version("2.1.0-dev1");
 	
 	public static InputHandler input; // input used in Game, Player, and just about all the *Menu classes.
 	public static Player player;
