+++ src/minicraft/level/tile/Wool/GreenWoolTile.java
@@ -1,4 +1,4 @@
-package minicraft.level.tile.WoolTiles;
+package minicraft.level.tile.Wool;
 
 import minicraft.core.io.Sound;
 import minicraft.entity.Direction;
@@ -13,12 +13,12 @@ import minicraft.level.Level;
 import minicraft.level.tile.Tile;
 import minicraft.level.tile.Tiles;
 
-public class YellowWoolTile extends Tile {
+public class GreenWoolTile extends Tile {
 
-    private static Sprite sprite = Sprite.repeat(17, 0, 2, 2, Color.get(550, 661, 440, 550));
+    private static Sprite sprite = Sprite.repeat(17, 0, 2, 2, Color.get(30, 40, 40, 50));
 
-    public YellowWoolTile() {
-        super("Yellow Wool", sprite);
+    public GreenWoolTile() {
+        super("Green Wool", sprite);
     }
 
     public boolean interact(Level level, int xt, int yt, Player player, Item item, Direction attackDir) {
@@ -27,7 +27,7 @@ public class YellowWoolTile extends Tile {
             if (tool.type == ToolType.Shovel) {
                 if (player.payStamina(3 - tool.level) && tool.payDurability()) {
                     level.setTile(xt, yt, Tiles.get("hole"));
-                    level.dropItem(xt*16+8, yt*16+8, Items.get("Yellow Wool"));
+                    level.dropItem(xt*16+8, yt*16+8, Items.get("Green Wool"));
                     Sound.monsterHurt.play();
                     return true;
                 }
