+++ src/minicraft/saveload/Load.java
@@ -287,7 +287,7 @@ public class Load {
 		
 		for (String keymap : subdata) {
 			String[] map = keymap.split(";");
-			if(prefVer.compareTo(new Version("2.0.6-dev1")) < 0)
+			if(prefVer.compareTo(new Version("2.1.0-dev1")) < 0)
 				map[0] = map[0].replaceAll("(UP|DOWN|LEFT|RIGHT)", "MOVE-\\1");
 			Game.input.setKey(map[0], map[1]);
 		}
@@ -329,7 +329,7 @@ public class Load {
 						if(Math.random() < 0.8) // don't replace *all* the lapis
 							tilename = "Gem Ore";
 					}
-					if(tilename.equalsIgnoreCase("WOOL") && worldVer.compareTo(new Version("2.0.6-dev1")) < 0) {
+					if(tilename.equalsIgnoreCase("WOOL") && worldVer.compareTo(new Version("2.1.0-dev1")) < 0) {
 						switch(Integer.parseInt(extradata.get(tileidx))) {
 							case 1:
 								tilename = "Red Wool";
@@ -475,7 +475,7 @@ public class Load {
 			name = name.replace("I.Armor", "Iron Armor").replace("S.Armor", "Snake Armor").replace("L.Armor", "Leather Armor").replace("G.Armor", "Gold Armor").replace("BrickWall", "Wall");
 		}
 		
-		if(worldVer.compareTo(new Version("2.0.6-dev1")) < 0)
+		if(worldVer.compareTo(new Version("2.1.0-dev1")) < 0)
 			name = name.replace("Book", "Editable Book");
 		
 		return name;
